+++ src/gui/wallet.go
@@ -314,6 +314,7 @@ func walletCreate(gateway Gatewayer) http.HandlerFunc {
 			Label:    label,
 			Encrypt:  encrypt,
 			Password: []byte(password),
+			ScanN:    scanN,
 		})
 		if err != nil {
 			switch err {
@@ -326,13 +327,6 @@ func walletCreate(gateway Gatewayer) http.HandlerFunc {
 			}
 		}
 
-		wlt, err = gateway.ScanAheadWalletAddresses(wlt.Filename(), []byte(password), scanN-1)
-		if err != nil {
-			logger.Errorf("gateway.ScanAheadWalletAddresses failed: %v", err)
-			wh.Error500(w)
-			return
-		}
-		// Wipes all sensitive data
 		rlt, err := NewWalletResponse(wlt)
 		if err != nil {
 			wh.Error500Msg(w, err.Error())
@@ -531,10 +525,15 @@ func walletTransactionsHandler(gateway Gatewayer) http.HandlerFunc {
 }
 
 // Returns all loaded wallets
-// RUI: /wallets
+// URI: /wallets
 // Method: GET
 func walletsHandler(gateway Gatewayer) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
+		if r.Method != http.MethodGet {
+			wh.Error405(w)
+			return
+		}
+
 		wlts, err := gateway.GetWallets()
 		if err != nil {
 			switch err {
@@ -546,7 +545,7 @@ func walletsHandler(gateway Gatewayer) http.HandlerFunc {
 			return
 		}
 
-		var wrs []*WalletResponse
+		wrs := make([]*WalletResponse, 0, len(wlts))
 		for _, wlt := range wlts {
 			wr, err := NewWalletResponse(wlt)
 			if err != nil {