+++ src/visor/rpc.go
@@ -32,11 +32,6 @@ func NewTransactionResult(tx *Transaction) (*TransactionResult, error) {
 	}, nil
 }
 
-// ReadableBlocks an array of readable blocks.
-type ReadableBlocks struct {
-	Blocks []ReadableBlock `json:"blocks"`
-}
-
 // TransactionResults array of transaction results
 type TransactionResults struct {
 	Txns []TransactionResult `json:"txns"`
@@ -137,8 +132,8 @@ func (rpc RPC) GetAddressTxns(v *Visor, addr cipher.Address) ([]Transaction, err
 }
 
 // CreateWallet creates new wallet
-func (rpc *RPC) CreateWallet(wltName string, options wallet.Options) (*wallet.Wallet, error) {
-	return rpc.v.Wallets.CreateWallet(wltName, options)
+func (rpc *RPC) CreateWallet(wltName string, options wallet.Options, bg wallet.BalanceGetter) (*wallet.Wallet, error) {
+	return rpc.v.Wallets.CreateWallet(wltName, options, bg)
 }
 
 // NewAddresses generates new addresses in given wallet